download release history
Release history contains the name, date and description of each release
at a given repo. These fields are fetched from the json object from
GitHub.

Intended workflow: After accepting a release, the user goes to GitHub
and fill in the description manually in release-history. After that the
download-release-history command is given.